# backend/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Abrolhos Blog API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /posts:
    get:
      operationId: listPublishedPosts
      tags: [Posts]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'

    post:
      operationId: createPost
      tags: [Posts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /posts/{slug}:
    get:
      operationId: getPublishedPostBySlug
      tags: [Posts]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'

        '404':
          description: Not Found
  /posts/{id}:
    put:
      operationId: updatePost
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
    delete:
      operationId: deletePost
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
  /posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /posts/{id}/archive:
    post:
      operationId: archivePost
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /admin/posts:
    get:
      operationId: adminListPosts
      tags: [Admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
  /categories:
    get:
      operationId: listCategories
      tags: [Categories]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'

  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [Categories]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

        '404':
          description: Not Found
  /admin/categories:
    post:
      operationId: createCategory
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /admin/categories/{id}:
    put:
      operationId: updateCategory
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Not Found
    delete:
      operationId: deleteCategory
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /tags:
    get:
      operationId: listTags
      tags: [Tags]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'

  /tags/{slug}:
    get:
      operationId: getTagBySlug
      tags: [Tags]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'

        '404':
          description: Not Found
  /categories/{slug}/posts:
    get:
      tags:
        - categories
      summary: Get posts by category
      operationId: getPostsByCategory
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'

        '404':
          description: Category not found
  /tags/{slug}/posts:
    get:
      tags:
        - tags
      summary: Get posts by tag
      operationId: getPostsByTag
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'

        '404':
          description: Tag not found
  /admin/tags:
    post:
      operationId: createTag
      tags: [Tags]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
  /admin/tags/{id}:
    put:
      operationId: updateTag
      tags: [Tags]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '404':
          description: Not Found
    delete:
      operationId: deleteTag
      tags: [Tags]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 255
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
        content:
          type: string
          minLength: 1
        categoryName:
          type: string
          maxLength: 100
          nullable: true
        tagNames:
          type: array
          items:
            type: string
            pattern: "^[a-z0-9_-]{1,50}$"
          nullable: true
    PostResponse:
      type: object
      required: [title, slug, content, status, authorUsername, categoryName, tagNames, createdAt, updatedAt]
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, SCHEDULED, ARCHIVED]
        authorUsername:
          type: string
        categoryName:
          type: string
        tagNames:
          type: array
          items:
            type: string
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9\s\-_]+$'
    CategoryResponse:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: ulid
        name:
          type: string
        slug:
          type: string
    TagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9\-_]+$'
    TagResponse:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: ulid
        name:
          type: string
        slug:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT